#8 Bruchmengen: Mengen-Engine von Integer auf Decimal (NUMERIC(14,3)) (#78)
Alle Bestandsmengen sind ab jetzt Dezimalzahlen – damit funktionieren
Abrechnung/Reservierung/FIFO für JEDE Einheit (kg, m², m³, l) inkl.
Bruchmengen, nicht mehr nur ganze Stück.

Backend
- Spalten instances.quantity/reserved_quantity, orders.quantity,
  order_lines.quantity, purchase_orders.quantity, sales.quantity →
  NUMERIC(14,3) (Migration 055 + idempotenter ALTER-Safety-Net in main.py).
- Neue EINE Mengen-Stelle services/quantity.py (to_qty/qty_sum/is_whole/qty_key):
  kapselt Umwandlung/Rundung (3 NK, exakt, kein float).
- reservation/inventory/subject/process/resource/recovery/scrap/serialization
  rechnen durchgängig mit Decimal. Die Reservierungs-Map (instances.reservations)
  speichert Mengen JSON-sicher als String; events.emit serialisiert Decimal→float.
- Einzelteil-Artikel (serialization='unit') nur GANZE Stück (Router-Check gegen
  die Serialisierung); Chargen (batch) tragen Bruchmengen. required_sample liefert
  ganze Proben (ceil, auch für 2.5 kg).
- API-Rand: Mengen als float (TS number – Frontend unverändert). AI-Tools geben
  Mengen als Zahl (kein int-Truncate mehr).

Frontend
- Menge/Ressourcen-Zeile/Teil-Verschrottung akzeptieren Nachkommastellen
  (kein Math.trunc mehr; step="any"/inputMode="decimal").

Tests: tests/test_fractional_quantities.py (Decimal-Roundtrip, FIFO, Stichprobe).

+++ backend/app/models/order.py
@@ -1,7 +1,8 @@
 from datetime import date, datetime
+from decimal import Decimal
 from typing import Optional
 
-from sqlalchemy import BigInteger, Boolean, Date, DateTime, Integer, String
+from sqlalchemy import BigInteger, Boolean, Date, DateTime, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from ..core.database import Base
@@ -36,7 +37,8 @@ class Order(Base, TimestampMixin):
     #             vorgewählte Instanzen → Subjekt FIFO ab Lager (Verkauf/Entnahme).
     #   chosen  – wirkt auf vorgewählte, vorhandene Instanzen (``instances.subject_of_order_id``).
     article_id: Mapped[Optional[int]] = mapped_column(BigInteger, index=True, nullable=True)
-    quantity: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
+    # Bestellmenge als Dezimalzahl (NUMERIC(14,3)) – ganze Stück ODER Bruchmenge (kg/m²/m³/l).
+    quantity: Mapped[Optional[Decimal]] = mapped_column(Numeric(14, 3), nullable=True)
     desired_delivery_date: Mapped[Optional[date]] = mapped_column(Date, nullable=True)
 
     # Stripe-Bezüge (Quelle der Wahrheit für Zahlung/Abo): die Checkout-Session, über die
