+++ backend/app/models/purchase_order.py
@@ -34,7 +34,8 @@ class PurchaseOrder(Base, TimestampMixin):
 
     order_id: Mapped[int] = mapped_column(BigInteger, index=True, nullable=False)
     article_id: Mapped[int] = mapped_column(BigInteger, index=True, nullable=False)
-    quantity: Mapped[int] = mapped_column(Integer, nullable=False)
+    # Bestellmenge als Dezimalzahl (NUMERIC(14,3)) – ganze Stück ODER Bruchmenge (kg/m²/…).
+    quantity: Mapped[Decimal] = mapped_column(Numeric(14, 3), nullable=False)
     # Routing: an welche Prozessschritt-Definition gebunden (mehrere möglich).
     step_id: Mapped[Optional[int]] = mapped_column(BigInteger, index=True, nullable=True)
